requirements.txt
asgiref==3.8.1
bidict==0.23.1
Django==4.2
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
eventlet==0.36.1
greenlet==3.0.3
h11==0.14.0
Markdown==3.6
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.2
simple-websocket==1.0.0
sqlparse==0.5.0
tzdata==2024.1
wsproto==1.2.0